+++ cipher-C++/fileLoaders.h
@@ -14,5 +14,5 @@ namespace fileLoaders {
 
 	std::unordered_map<std::string, int> loadMapInt(std::string fileName);
 
-	std::unordered_map<std::string, double> loadMapDouble(std::string fileName);
+	std::unordered_map<std::string, float> loadMapDouble(std::string fileName);
 }
